(from AES) Roll back v2listener.py to 43af03219 [ci-skip]
That's the v1.6.2, plus https://github.com/datawire/apro/pull/1716,
which is a fix that EPO cares about.

This should resolve several regressions in the 1.7.x series are
resolved by removing the ability to set `insecure.action` on a
per-`Host`-resource basis, which was an ability added in 1.7.0.  This
reverts to the pre-1.7.0 behavior of having one `Host`'s insecure
action "win" and be used for all `Host`s.

[ci-skip] because the gold files need updated.